backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "zafir-backend"
version = "0.1.0"
authors = [
  { name="Zafir", email="example@example.com" },
]
description = "Backend para el Sistema de Gestión de Cotizaciones Quirúrgicas"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "motor",
    "pydantic",
    "python-dotenv",
    "pdfplumber"
]
